correct_defers.
In my effort to clean up the code I had put the defers before checking
if there was an error opening the file in the first place. Im not
certain but this feels like it could cause problems if we then tried to
close that file with the defer.

Looking at other examples the SOP seems to be to defer after the error
check not before so i will ... defer to them on that.