removed_dotenv_depend_and_add_readme.
We only had a single dependency and all it was doing was reading a .env
file to load in an optional notes dir. This can just be done by getting
the user to set the env variable themselves.

So in the interest of having a dependency free program i removed it and
added instructions to the readme to set the env variable yourself.