temp_dir_home_dir.
While it didn't seem to be causing problems, it did worry me a little
that we were making temp files in the current dir. It hadn't happened
yet but it worries me that we will leave and commit a temp file to some
other repo that im working on.

So to combat this I have moved the working dir for the temp files to the notes dir.